+++ src/Qowaiv.Validation.DataAnnotations/AnnotatedProperty.cs
@@ -1,4 +1,4 @@
-﻿using Qowaiv.Validation.DataAnnotations.Refelection;
+﻿using Qowaiv.Validation.DataAnnotations.Reflection;
 using System.Reflection;
 
 namespace Qowaiv.Validation.DataAnnotations;
